+++ easyengine/__init__.py
@@ -1,9 +1,0 @@
-__version__ = '0.1.0'
-
-
-from .exceptions import (
-	EasyEngineException
-)
-from .base import (
-	EasyEngine
-)