+++ easyengine/base.py
@@ -1,6 +1,7 @@
 from easyengine.resources import (
 	Install,
 	Site,
+	Secure,
 	Stack
 )
 
@@ -13,8 +14,10 @@ class EasyEngine(object):
 		if port:
 			self._port   = int(port)
 		
+		self.clean   = Clean(self._host, **self.auth)
 		self.install = Install(self._host, **self.auth)
 		self.site    = Site(self._host, **self.auth)
+		self.secure  = Secure(self._host, **self.auth)
 		self.stack   = Stack(self._host, **self.auth)
 
 	@property