+++ easyengine/resources/__init__.py
@@ -1,3 +1,5 @@
+from .clean import Clean
 from .install import Install
+from .secure import Secure
 from .site import Site
 from .stack import Stack