+++ HISTORY.rst
@@ -2,5 +2,11 @@
 
 History
 -------
-0.0.1 (2016-09-4)
+0.1.2 (2016-09-5)
+++++++++++++++++++
+* Add: Secure methods
+* Add: Stack methods
+
+0.1.0 (2016-09-4)
+++++++++++++++++++
 * Initial commit